Move language/timezone RHs to core module
