--- indico/legacy/webinterface/rh/resetTimezone.py
@@ -1,44 +1,0 @@
-# This file is part of Indico.
-# Copyright (C) 2002 - 2017 European Organization for Nuclear Research (CERN).
-#
-# Indico is free software; you can redistribute it and/or
-# modify it under the terms of the GNU General Public License as
-# published by the Free Software Foundation; either version 3 of the
-# License, or (at your option) any later version.
-#
-# Indico is distributed in the hope that it will be useful, but
-# WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-# General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with Indico; if not, see <http://www.gnu.org/licenses/>.
-
-from __future__ import unicode_literals
-
-from flask import request, session
-from pytz import common_timezones_set
-
-from indico.core.config import Config
-from indico.legacy.webinterface.rh.base import RH
-
-
-class RHResetTZ(RH):
-    def _process(self):
-        mode = request.form['tz_mode']
-        tz = request.form.get('tz')
-        update_user = request.form.get('update_user') == '1'
-
-        if mode == 'local':
-            session.timezone = 'LOCAL'
-        elif mode == 'user' and session.user:
-            session.timezone = session.user.settings.get('timezone', Config.getInstance().getDefaultTimezone())
-        elif mode == 'custom' and tz in common_timezones_set:
-            session.timezone = tz
-
-        if update_user:
-            session.user.settings.set('force_timezone', mode != 'local')
-            if tz:
-                session.user.settings.set('timezone', tz)
-
-        return '', 204
