+++ indico/modules/bootstrap/templates/bootstrap.html
@@ -27,7 +27,7 @@
                                 {{- title -}}
                             </a>
                         {% else %}
-                            <a data-href="{{ url_for('misc.change_lang') }}"
+                            <a data-href="{{ url_for('core.change_lang') }}"
                                data-method="POST"
                                data-params='{{ {'lang': name}|tojson|forceescape }}'
                                data-reload-after
