+++ indico/modules/core/blueprint.py
@@ -16,9 +16,10 @@
 
 from __future__ import unicode_literals
 
-from indico.modules.core.controllers import RHSettings
+from indico.modules.core.controllers import RHChangeLanguage, RHChangeTimezone, RHSettings
 from indico.web.flask.util import redirect_view
 from indico.web.flask.wrappers import IndicoBlueprint
+
 
 _bp = IndicoBlueprint('core', __name__, template_folder='templates', virtual_template_folder='core')
 
@@ -26,6 +27,10 @@ _bp.add_url_rule('/admin/settings/', 'settings', RHSettings, methods=('GET', 'PO
 
 # TODO: replace with an actual admin dashboard at some point
 _bp.add_url_rule('/admin/', 'admin_dashboard', view_func=redirect_view('.settings'))
+
+# Global operations
+_bp.add_url_rule('/change-language', 'change_lang', RHChangeLanguage, methods=('POST',))
+_bp.add_url_rule('/change-timezone', 'change_tz', RHChangeTimezone, methods=('POST',))
 
 # Allow loadbalancers etc to easily check whether the service is alive
 _bp.add_url_rule('/ping', 'ping', lambda: ('', 204))
